Consolidate error files into sync index & error_type
